project1/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg2==2.9.9
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
